prepere project for hw

+++ src/services/user.service.ts
@@ -1,4 +1,10 @@
-import { IUser, IUserDTO } from "../interfaces/user.interface";
+import { StatusCodesEnum } from "../enums/status-codes.enum";
+import { ApiError } from "../errors/api.errors";
+import {
+    IUser,
+    IUserCreateDTO,
+    IUserUpdateDTO,
+} from "../interfaces/user.interface";
 import { userRepository } from "../repository/user.repository";
 
 class UserService {
@@ -6,20 +12,46 @@ class UserService {
         return userRepository.getAll();
     }
 
-    public getById(userId: string): Promise<IUserDTO> {
-        return userRepository.getById(userId);
+    public async getById(userId: string): Promise<IUser> {
+        const user = await userRepository.getById(userId);
+        if (!user) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+        return user;
     }
 
-    public create(user: IUserDTO): Promise<IUser> {
+    public create(user: IUserCreateDTO): Promise<IUser> {
         return userRepository.create(user);
     }
 
-    public update(userId: string, userData: IUserDTO): Promise<IUser> {
-        return userRepository.update(userId, userData);
+    public async update(
+        userId: string,
+        userData: IUserUpdateDTO,
+    ): Promise<IUser> {
+        const user = await userRepository.update(userId, userData);
+        if (!user) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+        return await userRepository.update(userId, userData);
     }
 
-    public delete(userId: string): Promise<IUser> {
-        return userRepository.delete(userId);
+    public async delete(userId: string): Promise<void> {
+        const user = await userRepository.getById(userId);
+        if (!user) {
+            throw new ApiError("User not found", StatusCodesEnum.NOT_FOUND);
+        }
+        await userRepository.delete(userId);
+    }
+
+    public async isEmailUnique(email: string): Promise<void> {
+        const user = await userRepository.getByEmail(email);
+
+        if (user) {
+            throw new ApiError(
+                "User is already exists",
+                StatusCodesEnum.BAD_REQUEST,
+            );
+        }
     }
 }
